requirements.txt
autopep8==1.5.5
click==7.1.2
Flask==1.1.2
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
matlabengineforpython===R2020b
pycodestyle==2.6.0
toml==0.10.2
Werkzeug==1.0.1
